Introducing separate handling for 400 Bad Request errors so that PHLAT can properly handle and read PLR's responses. This is in response to the PLR changes made to send a 400 code instead of 200 whenever the request message is invalid and results in an error acknowledgment.
Additional changes include changing Effective Start Date null checks to a StringUtils.hasText() check, renaming the Group Action map method in ProcessData so that it doesn't appear in the front-end as a modifiable field, and updating SQL scripts to increase the Property CHID columns from a length of 10 to a length of 15.

This version is confirmed to work for updating the Property fields (HDS Sub Type, PCN details, and Address Unit) and for HDS Name.